Extract the logic of managing SAM cloudformation stack from the the S… (#2740)
* Extract the logic of managing SAM cloudformation stack from the the S3-bucket bootstrap logic

* Move ManagedStackError class to the module 'managed_cloudformation_stack'
